# public/denis23x.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "description": "This API provides access to a collection of books, users, and reviews, allowing clients to perform various operations like viewing available books, creating and managing users, and posting or retrieving book reviews.\n\n- **Books**: Retrieve book details, search for books, and view books.\n- **Users**: Manage users, view user profiles.\n- **Reviews**: Add, view, update, or delete reviews for books.",
    "title": "Books Showcase Demo",
    "contact": {
      "email": "denix23x@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/api/v1/placeholder/reviews/{id}": {
      "get": {
        "summary": "Get Review",
        "deprecated": false,
        "description": "",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 615,
                        "userId": 388,
                        "bookId": 212,
                        "message": "Supra benigne statim correptius curo coniuratio tamquam cenaculum nihil.",
                        "rating": 1,
                        "createdAt": "2023-12-12T19:14:39.514Z",
                        "updatedAt": "2024-11-13T21:42:08.540Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete Review",
        "deprecated": false,
        "description": "",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 473,
                        "userId": 82,
                        "bookId": 360,
                        "message": "Despecto aveho ventus turba tabula desparatus acer.",
                        "rating": 2,
                        "createdAt": "2024-04-12T05:36:03.716Z",
                        "updatedAt": "2024-11-14T05:44:15.509Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update Review",
        "deprecated": false,
        "description": "",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "title": "Related User",
                    "description": "Related User",
                    "minimum": 1
                  },
                  "bookId": {
                    "type": "integer",
                    "title": "Related Book",
                    "description": "Related Book",
                    "minimum": 1
                  },
                  "message": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "integer",
                    "title": "Rating for the Book",
                    "description": "Rating for the Book",
                    "minimum": 1,
                    "maximum": 5
                  }
                }
              },
              "example": {
                "userId": 498,
                "bookId": 770,
                "message": "Commodi ante acquiro currus arbustum venia tener tamisium comparo.",
                "rating": 2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "data": {
                        "id": 689,
                        "userId": 870,
                        "bookId": 233,
                        "message": "Totidem vos avaritia civis.",
                        "rating": 2,
                        "createdAt": "2024-11-13T12:26:41.634Z",
                        "updatedAt": "2024-11-14T00:37:56.907Z"
                      },
                      "status": 200
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "issues": [
                          {
                            "code": "too_small",
                            "minimum": 2,
                            "type": "string",
                            "inclusive": true,
                            "exact": false,
                            "message": "String must contain at least 2 character(s)",
                            "path": [
                              "title"
                            ]
                          }
                        ],
                        "name": "ZodError"
                      },
                      "status": 400
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaClientKnownRequestError",
                        "code": "P2025",
                        "clientVersion": "5.22.0",
                        "meta": {
                          "modelName": "modelName",
                          "cause": "Record not found."
                        }
                      },
                      "status": 404
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {}
                    },
                    "status": {
                      "type": "integer"
                    }
                  }
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "error": {
                        "name": "PrismaError",
                        "clientVersion": "5.22.0"
                      },
                      "status": 500
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id",
            "minimum": 1
          },
          "userId": {
            "type": "integer",
            "title": "Related User",
            "description": "Related User",
            "minimum": 1
          },
          "bookId": {
            "type": "integer",
            "title": "Related Book",
            "description": "Related Book",
            "minimum": 1
          },
          "message": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "title": "Rating for the Book",
            "description": "Rating for the Book",
            "minimum": 1,
            "maximum": 5
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": false
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Develop Env"
    },
    {
      "url": "https://test.your-api-server.com",
      "description": "Testing Env"
    }
  ],
...
}